app/src/main/java/com/arisurya/jetpack/filmsapp/ui/detail/DetailMovieViewModel.kt
package com.arisurya.jetpack.filmsapp.ui.detail

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import com.arisurya.jetpack.filmsapp.data.FilmsRepository
import com.arisurya.jetpack.filmsapp.data.source.local.entity.FilmEntity


class DetailMovieViewModel(private val filmsRepository: FilmsRepository) : ViewModel() {
    private lateinit var movieId: String

    fun setSelectedMovie(movieId: String) {
        this.movieId = movieId
    }

    fun getMovie(): LiveData<FilmEntity> = filmsRepository.getDetailMovie(movieId.toInt())

}